Fix error: conversion to 'aesencrand::result_type' {aka 'long unsigned int'} from 'long long int' may change the sign of the result [-Werror=sign-conversion]

+++ link/.local/include/aesrand.hpp
@@ -27,7 +27,7 @@ DEF_URBG_CLASS(aesencrand, simd128, uint64_t)
 	s.u64vec += roundKey.u64vec;
 	const __m128i penultimate = _mm_aesenc_si128(s.i64vec, roundKey.i64vec);
 	const __m128i result = _mm_aesenc_si128(penultimate, roundKey.i64vec);
-	return result[0];
+	return static_cast<uint64_t>(result[0]);
 }
 
 DEF_URBG_CLASS(aesdecrand, simd128, uint64_t)
@@ -37,7 +37,7 @@ DEF_URBG_CLASS(aesdecrand, simd128, uint64_t)
 	//const __m128i penultimate = _mm_aesenc_si128(s.i64vec, roundKey.i64vec);
 	const __m128i penultimate = _mm_aesdec_si128(s.i64vec, roundKey.i64vec); // (SDW)
 	const __m128i result = _mm_aesdec_si128(penultimate, roundKey.i64vec);
-	return result[0];
+	return static_cast<uint64_t>(result[0]);
 }
 #else
 #warning "__AES__ not defined"
